// src/types/PlayerStats.ts
export type PlayerStats = {
  level: number;
  sex: Sex;
  alignment: Alignment;
  hp: number;
  mp: number;
  strength: number;
  magic: number;
  vitality: number;
  intelligence: number;
  speed: number;
  luck: number;
  clsRng: number;
  lngRng: number;
  spell: number;
  support: number;
  pDef: number;
  mDef: number;
  critical: number;
  criticalDefense: number;
  lbChance: number;
  lbPower: number;
  taChance: number;
  taPower: number;
  pursuitChance: number;
  pursuitPower: number;
  finalCriticalCorrection: number;
  lbCap: number;
  expertise: number;
  xp: number;
  soulPoints: number;
  digitalizePoints: number;
  digitalizeDuration: number;
  bethelPoints: number;
  macca: number;
  magnetite: number;
  attack: number;
  spin: number;
  shot: number;
  rush: number;
  HP_regen: number;
  MP_regen: number;
  HP_regenFlat: number;
  MP_regenFlat: number;
  weapon_affinity: WeaponAffinity;
  move_speed: number;
  currentHpPercent: number;
  moonPhase: number;
  timePeriod: "day" | "night";
  digitalizeActive: boolean;
  deityFamilyDmg: number;
  deityFamilyDmgTaken: number;
  vileFamilyDmg: number;
  vileFamilyDmgTaken: number;
  avianFamilyDmg: number;
  avianFamilyDmgTaken: number;
  megamiFamilyDmg: number;
  megamiFamilyDmgTaken: number;
  amatsuFamilyDmg: number;
  amatsuFamilyDmgTaken: number;
  raptorFamilyDmg: number;
  raptorFamilyDmgTaken: number;
  divineFamilyDmg: number;
  divineFamilyDmgTaken: number;
  jakiFamilyDmg: number;
  jakiFamilyDmgTaken: number;
  flightFamilyDmg: number;
  flightFamilyDmgTaken: number;
  yomaFamilyDmg: number;
  yomaFamilyDmgTaken: number;
  jiraeFamilyDmg: number;
  jiraeFamilyDmgTaken: number;
  machineFamilyDmg: number;
  machineFamilyDmgTaken: number;
  reaperFamilyDmg: number;
  reaperFamilyDmgTaken: number;
  holyFamilyDmg: number;
  holyFamilyDmgTaken: number;
  beastFamilyDmg: number;
  beastFamilyDmgTaken: number;
  fairyFamilyDmg: number;
  fairyFamilyDmgTaken: number;
  elementFamilyDmg: number;
  elementFamilyDmgTaken: number;
  fiendFamilyDmg: number;
  fiendFamilyDmgTaken: number;
  genmaFamilyDmg: number;
  genmaFamilyDmgTaken: number;
  wilderFamilyDmg: number;
  wilderFamilyDmgTaken: number;
  snakeFamilyDmg: number;
  snakeFamilyDmgTaken: number;
  nightFamilyDmg: number;
  nightFamilyDmgTaken: number;
  avatarFamilyDmg: number;
  avatarFamilyDmgTaken: number;
  foulFamilyDmg: number;
  foulFamilyDmgTaken: number;
  bruteFamilyDmg: number;
  bruteFamilyDmgTaken: number;
  hauntFamilyDmg: number;
  hauntFamilyDmgTaken: number;
  dragonFamilyDmg: number;
  dragonFamilyDmgTaken: number;
  fallenFamilyDmg: number;
  fallenFamilyDmgTaken: number;
  femmeFamilyDmg: number;
  femmeFamilyDmgTaken: number;
  kunitsuFamilyDmg: number;
  kunitsuFamilyDmgTaken: number;
  ladyFamilyDmg: number;
  ladyFamilyDmgTaken: number;
  drakeFamilyDmg: number;
  drakeFamilyDmgTaken: number;
  kishinFamilyDmg: number;
  kishinFamilyDmgTaken: number;
  omegaFamilyDmg: number;
  omegaFamilyDmgTaken: number;
  tyrantFamilyDmg: number;
  tyrantFamilyDmgTaken: number;
  heraldFamilyDmg: number;
  heraldFamilyDmgTaken: number;
  slashResist: number;
  bluntResist: number;
  thrustResist: number;
  handgunResist: number;
  penetrateResist: number;
  spreadResist: number;
  fireResist: number;
  iceResist: number;
  electricResist: number;
  forceResist: number;
  expelResist: number;
  deathResist: number;
  mysticResist: number;
  nerveResist: number;
  mindResist: number;
  almightyResist: number;
  slashBoost: number;
  bluntBoost: number;
  thrustBoost: number;
  handgunBoost: number;
  penetrateBoost: number;
  spreadBoost: number;
  fireBoost: number;
  iceBoost: number;
  electricBoost: number;
  forceBoost: number;
  expelBoost: number;
  deathBoost: number;
  mysticBoost: number;
  nerveBoost: number;
  mindBoost: number;
  almightyBoost: number;
  slashCap: number;
  thrustCap: number;
  bluntCap: number;
  handgunCap: number;
  spreadCap: number;
  penetrateCap: number;
  fireCap: number;
  iceCap: number;
  electricCap: number;
  forceCap: number;
  expelCap: number;
  deathCap: number;
  mysticCap: number;
  nerveCap: number;
  mindCap: number;
  almightyCap: number;
};